refactor - move unit tests for new endpoint to separate file
