+++ domain-cc/cc-app/tests/test_api.py
@@ -1,20 +1,8 @@
 from fastapi.testclient import TestClient
 
-TUBERCULOSIS_CLASSIFICATION = {
-    "diagnostic_code": 7710,
-    "classification_code": 6890,
-    "classification_name": "Tuberculosis",
-}
-BENIGN_GROWTH_BRAIN_CLASSIFICATION = {
-    "diagnostic_code": 8003,
-    "classification_code": 8964,
-    "classification_name": "Cyst/Benign Growth - Neurological other System",
-}
-DRUG_INDUCED_PULMONARY_PNEMONIA_CLASSIFICATION = {
-    "diagnostic_code": 6829,
-    "classification_code": 9012,
-    "classification_name": "Respiratory",
-}
+from .conftest import (BENIGN_GROWTH_BRAIN_CLASSIFICATION,
+                       DRUG_INDUCED_PULMONARY_PNEMONIA_CLASSIFICATION,
+                       TUBERCULOSIS_CLASSIFICATION)
 
 
 def test_classification(client: TestClient):
@@ -108,66 +96,3 @@ def test_v3_table_diagnostic_code(client: TestClient):
         response.json()["classification_name"]
         == DRUG_INDUCED_PULMONARY_PNEMONIA_CLASSIFICATION["classification_name"]
     )
-
-def test_diagnostic_code_mapping(client: TestClient):
-    """ classifier will fall back to dropdown lookup table if diagnostic code is not found """
-    json_post_dict = {
-        "diagnostic_code": TUBERCULOSIS_CLASSIFICATION["diagnostic_code"],
-        "claim_id": 100,
-        "form526_submission_id": 500,
-        "contention_text": "this_is_a_test",
-        "claim_type": "claim_for_increase",
-    }
-
-    response = client.post("/v2/classifier", json=json_post_dict)
-    assert response.status_code == 200
-    assert (
-            response.json()["classification_code"]
-            == TUBERCULOSIS_CLASSIFICATION["classification_code"]
-    )
-def test_classification_dropdown_cfi(client: TestClient):
-    """ classifier will fall back to dropdown lookup table if diagnostic code is not found """
-    json_post_dict = {
-        "diagnostic_code": 999999999,
-        "claim_id": 100,
-        "form526_submission_id": 500,
-        "contention_text": "Tuberculosis",
-        "claim_type": "claim_for_increase",
-    }
-
-    response = client.post("/v2/classifier", json=json_post_dict)
-    assert response.status_code == 200
-    assert (
-            response.json()["classification_code"]
-            == TUBERCULOSIS_CLASSIFICATION["classification_code"]
-    )
-
-
-def test_classification_dropdown_new(client: TestClient):
-    json_post_dict = {
-        "claim_id": 100,
-        "form526_submission_id": 500,
-        "contention_text": "Tuberculosis",
-        "claim_type": "new",
-    }
-
-    response = client.post("/v2/classifier", json=json_post_dict)
-    assert response.status_code == 200
-    assert (
-            response.json()["classification_code"]
-            == TUBERCULOSIS_CLASSIFICATION["classification_code"]
-    )
-
-def test_v2_null_response(client: TestClient):
-    json_post_dict = {
-        "diagnostic_code": 7,
-        "claim_id": 700,
-        "form526_submission_id": 777,
-        "claim_type": "claim_for_increase",
-        "contention_text": "this_is_a_test"
-    }
-
-    response = client.post("/v2/classifier", json=json_post_dict)
-    assert response.status_code == 200
-    assert response.json() is None
-
